refactor: give every element preview its own file
Previews were scattered: AudioPreview inline in results.tsx, Image and
Clip inline in registry.tsx, AnimatedImage in its own file. Each element
type now has a sister file under preview/, registry.tsx is only the
table, and results.tsx keeps just the shared primitives.

Also fixes an autosave test fixture missing the `pinned` field that
ElementSnapshot gained in #480.
